# docs/system-designer/02_OPENAPI_SPEC.yaml
openapi: 3.1.0
info:
  title: AgentLens API
  version: 2.1.0
  description: Global AI Agent, Harness & MCP News Curation Intelligence Platform RESTful API Specification
servers:
  - url: http://localhost:8000
    description: Local Backend Server
paths:
  /api/news:
    get:
      summary: 뉴스 피드 목록 다차원 필터링 및 페이징 조회
      tags: [News]
      parameters:
        - name: category
          in: query
          required: false
          description: 카테고리 필터 (all, harness, mcp_plugins_skills, agent_tech, ai_news)
          schema:
            type: string
            default: all
        - name: source
          in: query
          required: false
          description: 소스 코드 필터 (all, rss, rss_kr, github, arxiv, reddit, hackernews, huggingface)
          schema:
            type: string
            default: all
        - name: search
          in: query
          required: false
          description: 키워드 검색어 (제목, 요약, 태그, 기술스택, 시사점 통합 검색)
          schema:
            type: string
        - name: sort
          in: query
          required: false
          description: 정렬 기준 (hot, latest, top, quality)
          schema:
            type: string
            default: hot
        - name: high_signal_only
          in: query
          required: false
          description: ⭐️ Must-Read (High Signal) 아티클만 필터링 여부
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          required: false
          description: 페이지 번호 (1부터 시작)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: size
          in: query
          required: false
          description: 페이지 당 아이템 수
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: 뉴스 목록 페이징 결과 반환 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsListResponse'
...
  /api/news/{news_id}:
    get:
      summary: 특정 뉴스 아이템 단건 상세 조회
      tags: [News]
      parameters:
        - name: news_id
          in: path
          required: true
          description: 아티클 고유 ID
          schema:
            type: integer
      responses:
        '200':
          description: 아티클 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsItemResponse'
        '404':
          description: 해당 ID의 아티클을 찾을 수 없음
  /api/news/{news_id}/ask:
    post:
      summary: 특정 뉴스 아이템 대상 대화형 AI 질의응답 (Ask AI)
      tags: [News]
      parameters:
        - name: news_id
          in: path
          required: true
          description: 질문 대상 아티클 고유 ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskQuestionRequest'
      responses:
        '200':
          description: AI 답변 반환 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskQuestionResponse'
        '404':
          description: 아티클을 찾을 수 없음
  /api/news/{news_id}/ask/stream:
    post:
      summary: Ask AI 실시간 스트리밍 질의응답 (Server-Sent Events)
      tags: [News]
      parameters:
        - name: news_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskQuestionRequest'
      responses:
        '200':
          description: SSE 이벤트 스트림 (chunk, done, error)
          content:
            text/event-stream:
              schema:
                type: string
        '404':
          description: 아티클을 찾을 수 없음
  /api/news/feed/rss.xml:
    get:
      summary: RSS 2.0 XML 피드 발행
      tags: [News]
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: RSS 2.0 XML 문서 반환
          content:
            application/xml:
              schema:
                type: string
...
components:
  schemas:
...
    NewsItemResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Anthropic Claude 3.7 Sonnet & Hybrid Reasoning Model Released
        url:
          type: string
          example: https://www.anthropic.com/news/claude-3-7-sonnet
        source:
          type: string
          example: rss
        category:
          type: string
          example: ai_news
        summary:
          type: string
          example: Anthropic announces Claude 3.7 Sonnet featuring hybrid reasoning capabilities.
        author:
          type: string
          nullable: true
          example: Anthropic Research
        tags:
          type: array
          items:
            type: string
          example: [claude, reasoning, hybrid]
        tldr_bullets:
          type: array
          items:
            type: string
          example:
            - 하이브리드 추론 아키텍처 도입
            - 코딩 에이전트 성능 SWE-bench 70.3% 달성
            - 개발자가 사고 시간(Thinking Budget) 미세 조정 가능
        why_it_matters:
          type: string
          example: 실시간 상호작용과 심층 추론을 단일 모델에서 결합하여 에이전트 런타임 비용을 대폭 절감합니다.
        tech_stack:
          type: array
          items:
            type: string
          example: [Claude 3.7, SWE-bench, Python]
        quality_score:
          type: number
          format: float
          example: 9.4
        is_high_signal:
          type: boolean
          example: true
        raw_score:
          type: integer
          example: 1250
        comments_count:
          type: integer
          example: 340
        hotness_score:
          type: number
          format: float
          example: 8.9
        published_at:
          type: string
          format: date-time
          example: '2026-09-15T09:00:00Z'
        created_at:
          type: string
          format: date-time
          example: '2026-09-15T09:05:00Z'
        dedup_hash:
          type: string
          example: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
      required: [id, title, url, source, category, quality_score, published_at, dedup_hash]
    NewsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NewsItemResponse'
        total:
          type: integer
          example: 216
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 11
      required: [items, total, page, size, total_pages]
    AskQuestionRequest:
      type: object
      properties:
        question:
          type: string
          example: 이 논문/도구가 SWE-bench 평가 파이프라인에 어떻게 활용될 수 있나요?
      required: [question]
    AskQuestionResponse:
      type: object
      properties:
        news_id:
          type: integer
          example: 1
        question:
          type: string
          example: 이 도구의 활용 방안은 무엇인가요?
        answer:
          type: string
          example: 본 도구는 Docker 샌드박스 격리 환경에서 SWE-bench 인스턴스를 병렬 실행하도록 지원합니다.
        model_used:
          type: string
          nullable: true
          example: gemini-2.0-flash
      required: [news_id, question, answer]
...